adding test path for the external config. Fixing bug with json convert

+++ Public/Install-HuBotScript.ps1
@@ -39,9 +39,14 @@ function Install-HuBotScript
 
     $Config = Import-HuBotConfiguration -ConfigPath $ConfigPath
 
+    if (-not(Test-Path -Path $Config.BotExternalScriptsPath))
+    {
+        throw "Cannot find Hubot external config at $($Config.BotExternalScriptsPath)."
+    }
+
     Start-Process -FilePath npm -ArgumentList "install $($Name) --save" -Wait -NoNewWindow -WorkingDirectory $Config.BotPath
 
-    [System.Collections.ArrayList]$extenalScripts = Get-Content -Path $Config.BotExternalScriptsPath  | ConvertFrom-Json
+    [System.Collections.ArrayList]$extenalScripts = Get-Content -Path $Config.BotExternalScriptsPath -Raw | ConvertFrom-Json
 
     if ($extenalScripts -notcontains $NameInConfig)
     {
